% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDS.population.R
\docType{data}
\name{RDS.population}
\alias{RDS.population}
\title{The list that specifies the output of the \code{population} function.}
\format{
An object of class \code{list} of length 4.
}
\usage{
RDS.population
}
\description{
This list contains the output of the function \code{population} so that it
could be used in the sampling examples. Depending on the performance of the
computer it could take a considerable amount of time to produce this output.
}
\keyword{datasets}
