% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDS.samplingEx.R
\docType{data}
\name{RDS.samplingEx}
\alias{RDS.samplingEx}
\title{The dataframe that specifies the output of the \code{samplingEx} function.}
\format{
An object of class \code{rds.data.frame} (inherits from \code{data.frame}) with 300 rows and 8 columns.
}
\usage{
RDS.samplingEx
}
\description{
The function \code{samplingEx} may take considerable time to produce output,
depending on the computer's performance. For a quick reference to the expected
result, a saved version of the output is available.
}
\keyword{datasets}
