% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDS.samplingP.R
\docType{data}
\name{RDS.samplingP}
\alias{RDS.samplingP}
\title{The dataframe that specifies the output of the \code{samplingP} function.}
\format{
An object of class \code{rds.data.frame} (inherits from \code{data.frame}) with 1899 rows and 7 columns.
}
\usage{
RDS.samplingP
}
\description{
The function \code{samplingP} may take considerable time to produce output,
depending on the computer's performance. For a quick reference to the expected
result, a saved version of the output is available.
}
\keyword{datasets}
