% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{RDS Population Simulator}
\usage{
population (N = 1000, p.ties = 0.33, minVal = 0, maxVal = 300, zeros = 2, dis_type = "rexp",
skew = 0.05, pr = .33, pa = .33, atype_char = "NULL", atype_res = "NULL")
}
\arguments{
\item{N}{Population size. Should be a positive integer}

\item{p.ties}{Number of ties to be in the population. Should be a positive integer}

\item{minVal}{The minimum degree size an individual in the population can have. Should be a positive
integer}

\item{maxVal}{The maximum degree size an individual in the population can have. Should be a positive
integer}

\item{zeros}{The maximum number of zeros the population can have}

\item{dis_type}{Specify the distribution of the degrees. If need a uniformly distributed degree
distribution dis_type should be "runif". If need a normally distributed degree distribution dis_type
should be "rnorm". If need a right skewed distribution dis_type should be "rexp" and the skew parameter
should be a positive decimal value. In need a negative skewed distribution then the skew parameter
should be a negative decimal value.}

\item{skew}{needed when the distribution is left or right skewed. When selecting the value for the
skewness it is advisable to first observe the range of values given from the rexp function and to see
whether the maximum value of the distribution is close to the defined maxVal.}

\item{pr}{Proportion of individuals that has '1' as the response in the response variable.}

\item{pa}{Proportion of individuals that has '1' as their character in the indipendent variable.}

\item{atype_char}{How the independent variable is associated to other external factors. "NULL" when
needs to be randomly distributed. If needs to be associated with the network size use "net". Can
associate with network size only. Must note these abbreviations when need to associate with these
variables. "network size = net"}

\item{atype_res}{How the response variable is associated to other external factors."NULL" when needs
to be randomly distributed. If needs to be associated with both independent variable and network size
use "char * net". Can associate with network size and the independent variable. Must note these
abbreviations when need to associate with these variables. "network size = net","independent
dichotonoous variable = char"}
}
\description{
Generates RDS populations according to the user prefered variability. Do note the computer
capcities when using the function since it can affect the functionality of the methods used.
A population size of 10,000 can be produced with a RAM of 4GB without an issue. But if higher
population sizes are needed then higher RAM is needed. The produced population consists of the
degree size, a dichotonomous independent variable, a dichotonomous respond variable and the IDs of
the respondents in the network of each respond.
}
\examples{
population(N = 5000,p.ties = 0.6,minVal = 10,zeros = 0,pr = .5,pa = .1,atype_char = "net",atype_res = "net*char")
population(N = 10000,p.ties = 0.6,minVal = 10,zeros = 10,pr = .5,pa = .3,atype_char = "1/net",atype_res = "(1/net)*char")
population(N = 3000,p.ties = 0.6,minVal = 10,zeros = 0,pr = .5,pa = .4,atype_char = "2*net",atype_res = "net")
}
