% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingEx.R
\name{samplingEx}
\alias{samplingEx}
\title{Extracts sample from external populations}
\usage{
samplingEx (population, degree, net, seeds, coupon, waves = NULL, size = NULL, prob=NULL,
prob_var = NULL, prop_sam = NULL, prop_sam_var=NULL, showids = T, timeInterval = NULL)
}
\arguments{
\item{population}{.csv file containing the population details. First column should contain the id of the individual.}

\item{degree}{.csv file containing the network sizes. If it's not provided the degree size would the number of individuals
in the networks provided. If both the networks and the degrees are not provided, then it will randomly produce degrees for each resondent}

\item{net}{.csv file containging the network of each individual. Should contain the same format
of the output of the 'netsize table' given from the output of the inbuilt function 'population'. If not provided would
randomly assign individuals considering the degree.}

\item{seeds}{Number of seeds. Should be a positive integer}

\item{coupon}{Number of coupons. Should be a positive
integer}

\item{degree_column}{Column including degrees}

\item{waves}{Number of waves. Should be a positive integer}

\item{size}{Sample size. Should be a positive integer}

\item{prob}{Whether the seeds are selected probabilistically. If want to be selected randomly
then leave ot as 'NULL' If want to relate it to a variable add the variable name and the it
should be related. For example if the seeds selected should depend on the network size then 'network'
should be used. If it is inversly proportional then should enter as '1/network'.}

\item{prop_sam}{If the individuals from the network of an individual should be chosen randomly
use "NULL". If the selection should be done probabilistically state the variable and how it should
be related.}

\item{showids}{If TRUE displas the IDs beign recruited.}

\item{timeInterval}{Days between recruitment waves. zeroth wave is assumed to recruit at the present
day}
}
\description{
Extract samples from populations that the user inputs from an external source
without using a inbult function to simulate a population. Data should be in .csv format. Inputs are
recrited as that is in format of the output given in the inbuilt function 'population'
}
\examples{
samplingEx(population, degree, net, seeds = 10, coupon = 2, waves =8)
}
