% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingP.R
\name{samplingP}
\alias{samplingP}
\title{Extracts sample from generated populations}
\usage{
samplingP (population, seeds, coupon, waves = NULL, size = NULL, prob=NULL, prop_sam = NULL,
showids = T, timeInterval = NULL)
}
\arguments{
\item{population}{Population generated using 'SimRDS::population' function. The list output
should be fed as the value for this parameter.}

\item{seeds}{Number of seeds. Should be a positive integer}

\item{coupon}{Number of coupons. Should be a positive
integer}

\item{waves}{Number of waves. Should be a positive integer}

\item{size}{Sample size. Should be a positive integer}

\item{prob}{Whether the seeds are selected probabilistically. If want to be selected randomly
then leave ot as 'NULL' If want to relate it to a variable add the variable name and the it
should be related. For example if the seeds selected should depend on the network size then 'network'
should be used. If it is inversly proportional then should enter as '1/network'.}

\item{prop_sam}{If the individuals from the network of an individual should be chosen randomly
use "NULL". If the selection should be done probabilistically state the variable and how it should
be related.}

\item{showids}{If TRUE displas the IDs beign recruited.}

\item{timeInterval}{Days between recruitment waves. zeroth wave is assumed to recruit at the present
day}
}
\description{
Extract samples from the genrated populations using 'SimRDS::population' function.
Can directly use this output to te RDS estimates in RDS package
}
\examples{
samplingP(population, seeds = 10, coupon = 2, waves = 8, prob = '1/network', prob_sum = 'chara')
}
